split rc_api.h into rc_api_request.h, rc_api_runtime.h, and rc_api_user.h
