+++ src/rapi/rc_api_user.c
@@ -1,4 +1,4 @@
-#include "rc_api.h"
+#include "rc_api_user.h"
 #include "rc_api_common.h"
 
 #include <string.h>
