+++ test/rapi/test_rc_api_common.c
@@ -1,5 +1,7 @@
 #include "../rapi/rc_api_common.h"
 
+#include "rc_api_runtime.h" /* for rc_fetch_image */
+
 #include "rc_compat.h"
 
 #include "../test_framework.h"
